+++ src/OpenTelemetry.Api/Metrics/DoubleObserverMetricHandleSdk.cs
