+++ src/OpenTelemetry.Api/Metrics/Export/AggregationType.cs
